ADD命令をtemplate化
