+++ test/CMakeLists.txt
@@ -37,6 +37,7 @@ set(nask_parse_bnfc_SRCS
 set(nask_parse_front_SRCS
   ../src/front_end/front_config.cc
   ../src/front_end/front_inst.cc
+  ../src/front_end/front_inst_add.cc
   ../src/front_end/front_inst_jmp.cc
   ../src/front_end/front_inst_no_param.cc
   ../src/front_end/front_inst_pseudo.cc
